Simplify mydatetime implementation
No need for `m` to be an instance attribute. It's only ever used by the
custom `strftime` implementation, so we can just make it a local
variable in that function.

(Alternatively, if it needs to be exposed to other methods, it could be
a class attribute...but not sure that's necessary).